scripts/requirements-gemini-pdf.txt
pdf2image==1.16.3
Pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1
tenacity==8.2.3
